Update GridcoinAutoNodeDebian.sh
Should consider the GEO-IP round robin dns entry gridcoin.is.dopeshit.net as I update it regularly with both A and AAAA records. Also removing the remmed out IPv6 would be a smart idea , the ip stake is much more mature and iptables,ipfw or any firewall can handle ipv6 just the same as ipv4 and I have noticed as an ipv6 full node I get a vast number of spidered out connections via ipv6 it could benifit the network.

such as after the list of nodes a line such as 
+echo "-=-=-=- Adding Backup GEO-IP Based Single Entry For All Nodes-=-=-=-"
+echo "addnode=gridcoin.is.dopeshit.net" | sudo tee -a $confign
+echo "-=-=-=- Creating config-=-=-=-"

It could be #remmed out by those whom do not want to use it in the script or removed if users do not want it as any other node , but the wiki list I personally update every 3 months , the and record gridcoin.is.dopeshit.net every few weeks if not at least every time a new  node goes on the wiki.